+++ source/ProcessManager.Example.Orchestrations.Tests/Fixtures/ExampleConsumerAppManager.cs
@@ -74,23 +74,31 @@ public class ExampleConsumerAppManager : IAsyncDisposable
     /// Used by consumer app to configure Process Manager Message Client.</param>
     /// <param name="processManagerNotifyTopicResources">Process Manager Notify topic.
     /// Used by consumer app to configure Process Manager Message Client.</param>
+    /// <param name="brs021fmdStartTopicResources">BRS-021 Forward Metered Data Start Topic.
+    /// Used by consumer app to configure Process Manager Message Client.</param>
+    /// <param name="brs021fmdNotifyTopicResources">BRS-021 Forward Metered Data Notify Topic.
+    /// Used by consumer app to configure Process Manager Message Client.</param>
     /// <param name="ediTopicResources">EDI topic resources used by the app. Will be created if not provided.</param>
     /// <param name="processManagerApiUrl">Base URL of the Process Manager general API.</param>
     /// <param name="orchestrationsApiUrl">Base URL of the Orchestrations API.</param>
     public async Task StartAsync(
         TopicResource processManagerStartTopicResources,
         TopicResource processManagerNotifyTopicResources,
+        TopicResource brs021fmdStartTopicResources,
+        TopicResource brs021fmdNotifyTopicResources,
         EdiTopicResources? ediTopicResources,
         string processManagerApiUrl,
         string orchestrationsApiUrl)
     {
-        ediTopicResources ??= await EdiTopicResources.CreateNew(ServiceBusResourceProvider);
+        ediTopicResources ??= await EdiTopicResources.CreateNewAsync(ServiceBusResourceProvider);
 
         // Prepare host settings
         var appHostSettings = CreateAppHostSettings(
             "ProcessManager.Example.Consumer",
             processManagerStartTopicResources,
             processManagerNotifyTopicResources,
+            brs021fmdStartTopicResources,
+            brs021fmdNotifyTopicResources,
             ediTopicResources,
             processManagerApiUrl,
             orchestrationsApiUrl);
@@ -157,6 +165,8 @@ public class ExampleConsumerAppManager : IAsyncDisposable
         string csprojName,
         TopicResource processManagerStartTopicResources,
         TopicResource processManagerNotifyTopicResources,
+        TopicResource brs021fmdStartTopicResources,
+        TopicResource brs021fmdNotifyTopicResources,
         EdiTopicResources ediTopicResources,
         string processManagerGeneralApiBaseUrl,
         string orchestrationApiBaseUrl)
@@ -214,6 +224,12 @@ public class ExampleConsumerAppManager : IAsyncDisposable
         appHostSettings.ProcessEnvironmentVariables.Add(
             $"{ProcessManagerServiceBusClientOptions.SectionName}__{nameof(ProcessManagerServiceBusClientOptions.NotifyTopicName)}",
             processManagerNotifyTopicResources.Name);
+        appHostSettings.ProcessEnvironmentVariables.Add(
+            $"{ProcessManagerServiceBusClientOptions.SectionName}__{nameof(ProcessManagerServiceBusClientOptions.Brs021ForwardMeteredDataStartTopicName)}",
+            brs021fmdStartTopicResources.Name);
+        appHostSettings.ProcessEnvironmentVariables.Add(
+            $"{ProcessManagerServiceBusClientOptions.SectionName}__{nameof(ProcessManagerServiceBusClientOptions.Brs021ForwardMeteredDataNotifyTopicName)}",
+            brs021fmdNotifyTopicResources.Name);
 
         // => Edi topic
         appHostSettings.ProcessEnvironmentVariables.Add(
@@ -237,7 +253,7 @@ public class ExampleConsumerAppManager : IAsyncDisposable
     {
         private const string EnqueueBrsX03SubscriptionName = $"enqueue-brs-x03-subscription";
 
-        public static async Task<EdiTopicResources> CreateNew(ServiceBusResourceProvider serviceBusResourceProvider)
+        public static async Task<EdiTopicResources> CreateNewAsync(ServiceBusResourceProvider serviceBusResourceProvider)
         {
             var ediTopicBuilder = serviceBusResourceProvider.BuildTopic("edi-topic");
             AddSubscriptionsToTopicBuilder(ediTopicBuilder);