+++ docs/ProcessManager.Components.Abstractions/ReleaseNotes/ReleaseNotes.md
@@ -1,5 +1,9 @@
 # ProcessManager.Components.Abstractions Release Notes
 
+## Version 1.3.6
+
+- Update NuGet packages.
+
 ## Version 1.3.5
 
 - Update NuGet packages.